+++ index.php
@@ -6,7 +6,7 @@
  * @author Oleg Grigoriev <go.vasac@gmail.com>
  * @license https://raw.github.com/axypro/errors/master/LICENSE MIT
  * @link https://github.com/axypro/errors repository
- * @link https://github.com/axypro/errors/blob/master/doc/README.md documentation
+ * @link https://github.com/axypro/errors/blob/master/README.md documentation
  * @link https://packagist.org/packages/axy/errors composer
  * @uses PHP5.4+
  */