+++ tests/tst/thrower/Thrower.php
@@ -5,15 +5,23 @@
 
 namespace axy\errors\tests\tst\thrower;
 
-use \axy\errors\tests\tst\errors\InvalidConfig;
+use axy\errors\tests\tst\errors\InvalidConfig;
 
 class Thrower
 {
+    /**
+     * @param string $ns
+     * @throws \axy\errors\tests\tst\errors\InvalidConfig
+     */
     public function run($ns)
     {
         $this->raise($ns);
     }
 
+    /**
+     * @param string $ns
+     * @throws \axy\errors\tests\tst\errors\InvalidConfig
+     */
     private function raise($ns)
     {
         if ($ns === true) {