+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## dev-master
 
-* Travis
+* Travis & Coveralls
 * Fix tests for PHP7 and native trace truncate
 
 ## 1.0.4 (13.11.2015)